fix(test): match production UID type in incident audit regression fixture
Use uint32(os.Geteuid()) to match productionRuntime.requiredOwnerUID.
The failed run stopped at test compilation before any production SSH.
No production logic changes; retain full native and PostgreSQL test gate.

+++ apps/api-go/internal/appcli/incident343_post_stop_linux_test.go
@@ -75,7 +75,7 @@ func TestIncident343ForwardAuditRejectsAdditionalProgress(t *testing.T) {
 			if err := os.WriteFile(path, []byte("{}\n"), 0600); err != nil {
 				t.Fatal(err)
 			}
-			runtime := &productionRuntime{requiredOwnerUID: os.Geteuid()}
+			runtime := &productionRuntime{requiredOwnerUID: uint32(os.Geteuid())}
 			if err := runtime.requireIncident343Entries(directory, []string{"before-status.json"}); err != nil {
 				t.Fatal(err)
 			}
